# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Segura de E-commerce",
        "description": "Documentación oficial de la API de Comercio Electrónico desarrollada con Laravel 12",
        "contact": {
            "email": "tu-email@ejemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor Local de la API"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registro de nuevos clientes",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan@ejemplo.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secreto123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "secreto123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado con éxito"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Inicio de sesión",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "juan@ejemplo.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secreto123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sesión iniciada correctamente"
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cierre de sesión",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listado público del catálogo de productos",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "responses": {
                    "201": {
                        "description": "Producto creado con éxito"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/{product}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Ver el detalle de un producto específico",
                "operationId": "960e17d05ed63211b20fb4b8505f1263",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del producto"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Editar un producto existente",
                "operationId": "d8f8f68cbee2f6a5c291d7940a3dc4ed",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto actualizado con éxito"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto",
                "operationId": "13d9c94a21e9746d2a61b6a3247b7693",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado correctamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticación",
            "description": "Autenticación"
        },
        {
            "name": "Productos",
            "description": "Productos"
        }
    ]
}